Added Material UI inputs and sort

+++ src/components/Hero/Table/Table.jsx
@@ -1,8 +1,9 @@
 import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
-import Pagination from '../../../pagination';
+import Pagination from '../Pagination/Pagination';
 import { deleteEmployee } from '../../../store/EmployeeContext/EmployeeSlice';
 import EditModal from '../../EditModal/EditModal';
+import { confirm } from "react-confirm-box";
 import "./Table.scss";
 
 function Table(props) { 
@@ -14,10 +15,82 @@ function Table(props) {
     setId(elId);
   }
 
-  const deleteHandler = (id) => {
-    dispatch(
-      deleteEmployee({id : id})
-    )
+  const deleteHandler = async (id) => {
+    const result = await confirm("Are you sure?");
+    if (result) {
+      dispatch(
+        deleteEmployee({id : id})
+      )
+      return;
+    }
+  };
+
+  let [count, setCount] = useState(0);
+  const nameHandler = () => {
+    if(count === 0){
+      let temp = [...props.search].sort((a, b) => 
+        b.name > a.name ? 1 : -1
+      );
+      props.setSearch(temp);
+      setCount(1)
+    }else{
+      let temp = [...props.search].sort((a, b) => 
+        b.name < a.name ? 1 : -1
+      );
+      props.setSearch(temp);
+      setCount(0)
+    }
+  }
+
+  let [email, setEmail] = useState(0);
+  const emailHandler = () => {
+    if(email === 0){
+      let temp = [...props.search].sort((a, b) => 
+        b.email > a.email ? 1 : -1
+      );
+      props.setSearch(temp);
+      setEmail(1)
+    }else{
+      let temp = [...props.search].sort((a, b) => 
+        b.email < a.email ? 1 : -1
+      );
+      props.setSearch(temp);
+      setEmail(0)
+    }
+  }
+
+  let [phone, setPhone] = useState(0);
+  const numberHandler = () => {
+    if(phone === 0){
+      let temp = [...props.search].sort((a, b) => 
+        b.number > a.number ? 1 : -1
+      );
+      props.setSearch(temp);
+      setPhone(1)
+    }else{
+      let temp = [...props.search].sort((a, b) => 
+        b.number < a.number ? 1 : -1
+      );
+      props.setSearch(temp);
+      setPhone(0)
+    }
+  }
+
+  let [department, setDepartment] = useState(0);
+  const departmentHandler = () => {
+    if(department === 0){
+      let temp = [...props.search].sort((a, b) => 
+        b.department > a.department ? 1 : -1
+      );
+      props.setSearch(temp);
+      setDepartment(1)
+    }else{
+      let temp = [...props.search].sort((a, b) => 
+        b.department < a.department ? 1 : -1
+      );
+      props.setSearch(temp);
+      setDepartment(0)
+    }
   }
 
   return (
@@ -25,25 +98,25 @@ function Table(props) {
       <table className="table">
         <thead>
           <tr>
-            <th scope="col">
+            <th scope="col" onClick={nameHandler}>
               Employee Name
               <button className="head-btn">
                 <i className="bx bx-up-arrow-alt arrow-btn" />
               </button>
             </th>
-            <th scope="col">
+            <th scope="col" onClick={emailHandler}>
               Email Address(Personal)
               <button className="head-btn">
                 <i className="bx bx-up-arrow-alt arrow-btn" />
               </button>
             </th>
-            <th scope="col" >
+            <th scope="col" onClick={numberHandler}>
               Mobile number
               <button className="head-btn">
                 <i className="bx bx-up-arrow-alt arrow-btn" />
               </button>
             </th>
-            <th scope="col">
+            <th scope="col" onClick={departmentHandler}>
               Department
               <button className="head-btn">
                 <i className="bx bx-up-arrow-alt arrow-btn" />
